db-loader/pyproject.toml
[project]
name = "db-loader"
version = "0.1.0"
description = "CIOOS Explorer db-loader service."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy<2.0.0",
    "pandas<2.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sentry-sdk>=2.38.0",
    "sqlalchemy>=2.0.0,<3.0.0",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cde_db_loader"]